tests: move PodStatus test to k8sconv
Missed the test when moving the logic, so just moving it to where
the actual code is.
