+++ internal/engine/k8swatch/pod_watch_test.go
@@ -2,7 +2,6 @@ package k8swatch
 
 import (
 	"context"
-	"fmt"
 	"strconv"
 	"sync"
 	"testing"
@@ -250,55 +249,6 @@ func TestPodWatchReadd(t *testing.T) {
 	// redeployed, we still need to broadcast the pod to make
 	// sure it gets repopulated.
 	f.assertObservedPods(p)
-}
-
-type podStatusTestCase struct {
-	pod      corev1.PodStatus
-	status   string
-	messages []string
-}
-
-func TestPodStatus(t *testing.T) {
-	cases := []podStatusTestCase{
-		{
-			pod: corev1.PodStatus{
-				ContainerStatuses: []corev1.ContainerStatus{
-					{
-						LastTerminationState: corev1.ContainerState{
-							Terminated: &corev1.ContainerStateTerminated{
-								ExitCode: 128,
-								Message:  "failed to create containerd task: OCI runtime create failed: container_linux.go:345: starting container process caused \"exec: \\\"/hello\\\": stat /hello: no such file or directory\": unknown",
-								Reason:   "StartError",
-							},
-						},
-						Ready: false,
-						State: corev1.ContainerState{
-							Waiting: &corev1.ContainerStateWaiting{
-								Message: "Back-off 40s restarting failed container=my-app pod=my-app-7bb79c789d-8h6n9_default(31369f71-df65-4352-b6bd-6d704a862699)",
-								Reason:  "CrashLoopBackOff",
-							},
-						},
-					},
-				},
-			},
-			status: "CrashLoopBackOff",
-			messages: []string{
-				"failed to create containerd task: OCI runtime create failed: container_linux.go:345: starting container process caused \"exec: \\\"/hello\\\": stat /hello: no such file or directory\": unknown",
-				"Back-off 40s restarting failed container=my-app pod=my-app-7bb79c789d-8h6n9_default(31369f71-df65-4352-b6bd-6d704a862699)",
-			},
-		},
-	}
-
-	for i, c := range cases {
-		t.Run(fmt.Sprintf("case%d", i), func(t *testing.T) {
-			pod := corev1.Pod{Status: c.pod}
-			status := k8sconv.PodStatusToString(pod)
-			assert.Equal(t, c.status, status)
-
-			messages := k8sconv.PodStatusErrorMessages(pod)
-			assert.Equal(t, c.messages, messages)
-		})
-	}
 }
 
 func (f *pwFixture) addManifestWithSelectors(manifestName string, ls ...labels.Selector) model.Manifest {
